+++ seleniumtid/pageelements/page_element.py
@@ -13,6 +13,8 @@ been supplied.
 
 from selenium.webdriver.remote.webelement import WebElement
 
+from seleniumtid.visual_test import VisualTest
+
 
 class PageElement(object):
     """
@@ -74,3 +76,13 @@ class PageElement(object):
         :returns: web element if it is not visible or False
         """
         return self.utils.wait_until_element_not_visible(self.locator, timeout)
+
+    def assertScreenshot(self, filename, threshold=0, exclude_elements=[]):
+        """Assert that a screenshot of the element is the same as a screenshot on disk, within a given threshold.
+
+        :param filename: the filename for the screenshot, which will be appended with ``.png``
+        :param threshold: the threshold for triggering a test failure
+        :param exclude_elements: list of CSS/XPATH selectors as a string or WebElement objects that must be excluded
+                                 from the assertion.
+        """
+        VisualTest().assertScreenshot(self.element(), filename, self.__class__.__name__, threshold, exclude_elements)